# Conflicts:
#	Core/Group.cs
#	Core/Student.cs
#	Core/Teacher.cs
#	Core/project.json
#	Infrastructure/BaseContext.cs
#	Infrastructure/Migrations/20170107125846_init1.cs
#	Infrastructure/Migrations/20170107130036_init2.cs
#	Infrastructure/Migrations/BaseContextModelSnapshot.cs
#	Infrastructure/StudentRepository.cs
#	Test/ControllerTester.cs
#	Test/project.lock.json
#	Web/Controllers/AbstractController.cs
#	Web/Controllers/CoursesController.cs
#	Web/Controllers/DataController.cs
#	Web/Controllers/StudentsController.cs
#	Web/Controllers/TeachersController.cs
#	Web/DummyData/Data.cs
#	Web/project.json

+++ Web/Controllers/AbstractController.cs
@@ -41,28 +41,21 @@ namespace Web.Controllers
                 return NotFound("Id " + id + " does not exist");
 
             entity.ApplyTo(entityForUpdate, ModelState);
- 
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
 
-            try
-            {
-                Repository.Update(entityForUpdate);
-            }
-            catch (Exception)
-            {
-                return BadRequest("Invalid operation(s). Verify if operation(s) comply with the requirements.");
-            }
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            Repository.Update(entityForUpdate);
             return Ok(entityForUpdate);
         }
 
         [HttpDelete("{id}")]
-        public void Delete(Guid id)
+        public IActionResult Delete(Guid id)
         {
             var entity = Repository.GetById(id);
+            if (entity == null)
+                return BadRequest("Id " + id + " doesn't exist");
             Repository.Delete(entity);
+            return NoContent();
         }
     }
 }
